consolidate movies compnenets
